moved phaser game canvas into its own component
