+++ phaser-react/src/pages/About.jsx
@@ -1,0 +1,7 @@
+//import React from 'react'
+
+export default function About() {
+  return (
+    <div>About</div>
+  )
+}
